DragNDropGenerator now can run an apple script when making a package.
While the DragNDropGenerator supports custom DS_Store and backgrounds, it is
still very hard to automatically setup nice looking packages. The primary
issue is that the DS_Store embeds the name of the volume in the path
to backgrounds, which means that if a package embeds the version in its
volume name a new DS_Store must generated for each release.

Instead you now can use applescript to setup your DS_Store.

This change also ensures that temporary RW image has enough space for these
changes, creating 1 MB dummy padding file, that is later removed from the image.

Co-Author: Adam Strzelecki <adam.strzelecki@java.pl>