DragNDropGenerator now manually computes the size of the temp dmg image.
When creating a DragNDrop bundle with a custom layout and background you need
to make sure that the image size is large enough to store both the background
and the custom .DS_Store.

Now previously the DragNDropGenerator relied on the fact that both of these
files would exist in the staging folder so it could let hdiutil compute an
exact size. Now that we allow generation of the .DS_Store to happen after
creation of the image we need to manually specify the size of the image
to include room for the new .DS_Store and background.

The size calculation must not take only file sizes into account, but also file
system block padding and directory entry structure size, therefore we add
additional 4K to each file size to provide more accurate final value.

The whole reason why we have to delay the background image copy, is that
if we let hdiutil package the background image it gets modified in such away
that applescript aren't able to specify it as the background image.