Revert "cmQtAutoGenerators: Fix rcc invocation for Qt 5.0 and 5.1 (#15644)"
This reverts commit 9a271e13236d81b79e3e367a6898e09d3dcf28d0.

It was the wrong thing to do, as made obvious in the bug report because it
changes cmake from doing the right thing (using --list) to doing the wrong
thing (using -list, which exists only for compatibility).

Qt 5.1 is old enough that cmake should not get new support for it in
CMake 3.4.  That would be a maintainance cost forever, including into Qt 6
times because of the use of the compatibility-only switch.