Revert "Revert "cmQtAutoGenerators: Fix rcc invocation for Qt 5.0 and 5.1 (#15644)""
This reverts commit 08147a9a0cee722196ddf8d63c25c5d1f8ebd336.
Discussion in the bug report pointed out that the proper fix is
to adapt based on the availability of -list or --list from the
version of rcc found.  Using -list is how to support all existing
versions of Qt.  Detecting when --list is available is how to
support new versions of Qt that offer it and future versions of
Qt that do not offer -list.